src/login-service/requirements.txt
flask
flask_cors
sqlalchemy
pyjwt
psycopg2
